Fixed dynamic directory path calculation for nil value.
Due to nature of how `Hash#fetch` works, a key value of `nil` would
prevent the `#fetch` falling back to `paths`. The use of the `||`
prevents this from happening so `paths` ends up as the proper fallback
value.

This also ensures `#split` doesn't attempt to split on a `nil` value
which would cause an exception.

Finally, there is a subtly with how `ENV` works. It behaves like a
`Hash` but is an `Object` which means some behavior is not consistent.
Namely:

    Hash.new[nil] # => nil
    ENV[nil] # => ERROR: No implicit conversion to string.

One way to test this would be to use `ENV.clear` but that contaminates
the entire test suit due to `ENV` being a singleton. Instead an empty
hash is used for the test but isn't a complete test.